Create user routes and controller.
