+++ src/app.ts
@@ -2,14 +2,14 @@ import express, { Express } from "express";
 import morgan from "morgan";
 import { Server } from "http";
 
-import userRoutes from "./api/v1/routes/userRoutes";
+import loanRoutes from "./api/v1/routes/loanRoutes";
 
 const app: Express = express();
 
 app.use(morgan("combined"));
 app.use(express.json());
 
-app.use("/api/v1/loans", userRoutes);
+app.use("/api/v1/loans", loanRoutes);
 
 const PORT: string | number = process.env.PORT || 3000;
 const server: Server = app.listen(PORT, () => {
